requirements.txt
aniso8601==9.0.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-RESTful==0.3.9
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.26
pkg-resources==0.0.0
protobuf==3.18.0
pytz==2021.1
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.5.0
